+++ app/src/main/java/eu/kanade/domain/source/service/SourcePreferences.kt
@@ -51,6 +51,11 @@ class SourcePreferences(
         emptySet(),
     )
 
+    fun invalidMangaExtensions() = preferenceStore.getStringSet(
+        Preference.appStateKey("invalid_manga_extensions"),
+        emptySet(),
+    )
+
     fun globalSearchFilterState() = preferenceStore.getBoolean(
         Preference.appStateKey("has_filters_toggle_state"),
         false,